src-python/requirements.txt
fastapi
websockets
uvicorn
pydantic>=2
pydantic-settings>=2
python-multipart
rich
dotenv
shortuuid

# Llama Index for RAG
llama-index
llama-index-llms-ollama
llama-index-llms-openai-like
llama-index-embeddings-ollama
llama-index-embeddings-openai-like
llama-index-llms-gemini
llama-index-vector-stores-chroma
chromadb
llama-index-tools-duckduckgo
json-repair
llama-index-postprocessor-colbert-rerank
llama-index-storage-docstore-mongodb

# observability
arize-phoenix
arize-phoenix-otel
openinference-instrumentation-llama-index

# typescript type generation
pydantic-to-typescript>=2
json2ts

# utils
black
pylint>=3.3.4
pyinstaller